Move all dynamic form-related code to new module: dynamic-forms
