+++ client/app/dynamic-forms/controls/abstract-form-control.ts
@@ -1,7 +1,7 @@
 import { Input } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 
-import { FormControlSpec } from '../form-control-spec.interface';
+import { FormControlSpec } from '../form-control-spec';
 
 /**
  * Base class for all dynamic reactive form controls.
