+++ client/app/dynamic-forms/controls/datetime-control.component.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
-import { AbstractFormControl } from './abstract-form-control.class';
+import { AbstractFormControl } from './abstract-form-control';
 import { DatetimeInputComponent } from '../../core/datetime-input/datetime-input.component';
 
 /**
